refactor: TicketingClient 및 TicketingResponse 패키지 이동
